+++ spec/integrations/pro/consumption/lrj_with_auto_offset_management.rb
@@ -8,7 +8,7 @@ setup_karafka do |config|
   config.license.token = pro_license_token
 end
 
-class Consumer < Karafka::BaseConsumer
+class Consumer < Karafka::Pro::BaseConsumer
   def consume
     DataCollector[0] << messages.last.offset
   end