+++ spec/integrations/pro/rails/rails5/railtie_setup.rb
@@ -22,7 +22,7 @@ setup_karafka do |config|
   config.license.token = pro_license_token
 end
 
-class Consumer < Karafka::BaseConsumer
+class Consumer < Karafka::Pro::BaseConsumer
   def consume
     DataCollector[0] << true
   end